requirements.txt
pydantic>=2.0.0
requests>=2.31.0
openai>=1.0.0
tavily-python>=0.3.0
python-dotenv>=1.0.0
pandas>=2.0.0
termcolor>=2.0.0
numpy>=1.24.0

